refactor step 1: move Selection and related classes to their own file
