+++ src/genjax/_src/generative_functions/combinators/vector/map_combinator.py
@@ -29,10 +29,12 @@ from genjax._src.core.datatypes.generative import (
     EmptyChoice,
     GenerativeFunction,
     HierarchicalChoiceMap,
-    HierarchicalSelection,
     JAXGenerativeFunction,
     Trace,
     TraceSlice,
+)
+from genjax._src.core.datatypes.selection import (
+    HierarchicalSelection,
 )
 from genjax._src.core.interpreters.incremental import Diff
 from genjax._src.core.pytree import Pytree
