+++ src/genjax/_src/generative_functions/combinators/vector/unfold_combinator.py
@@ -27,12 +27,12 @@ from genjax._src.core.datatypes.generative import (
     ChoiceMap,
     EmptyChoice,
     GenerativeFunction,
-    HierarchicalSelection,
     JAXGenerativeFunction,
     Mask,
     Trace,
     TraceSlice,
 )
+from genjax._src.core.datatypes.selection import HierarchicalSelection
 from genjax._src.core.interpreters.incremental import Diff
 from genjax._src.core.pytree import Pytree
 from genjax._src.core.typing import (
